Update build.yml
We've made big updates to the build.yml reference example for Terraform providers.

The most critical change is the new
`upload-terraform-registry-manifest-artifact` job. This is necessary for correctness of publishing providers to Terraform Registry.

I open this pull request as a convenient way of notifying your team of this change. In terraform-provider-cloudinit and terraform-provider-null, I have completely replaced the existing build.yml with this new version -- that is the simplest upgrade path.

Review and change the defaults as needed. Follow the TODO comments for step-by-step guidance.